Group Java compatibility tests
